New test bug/r300-readcache
This tests against a particular problem of the r300 driver, related to a
readcache coherency problem.

Note that while I had first found this problem via glean/texCube, the fact
that I did find the problem there was pure chance, because the bug depends
on the exact coordinate of the glReadPixels call (and thus on the window
position). This new test loops through many coordinates to make sure the
window coordinates are irrelevant.