Generate abbreviated results
The results files can get rather huge when tests fail, because tests like
glean/blendFunc output a lot of debugging data, which we all record.

Now, we generate an additional .../summary file, in which the info string
is simply truncated. Pretty stupid, but it should give enough info to get
an idea of the rough kind of failure.

Add a new option for piglit-summary-html.py, to choose between full or
abbreviated info when both are present.