glsl-lod-bias: Lots of changes
Previously the test had a pretty significant flaw that limited its
utility.  The texture was much smaller than the squares being drawn,
and the checkerboard tiles in the texture were 1x1.  This resulted in
the unbiased LOD being negative (magnification).  If any positive LOD
were selected, the result would be uniformly blurred tile colors.

This changes the test in serveral ways.

1. Make the texture larger and draw squares the same size.  This
results in an unbiased LOD of 0.0.

2. Pass the bias to the shader via a uniform.

3. Draw with all the positive LODs that don't cause checkerboard tiles
to be blurred into their neighbors.

4. Set the blend mode so that we actually get gray where the nominally
pink checkerboard tiles are drawn.