Add new depth-tex-modes-rg test.
Tests using GL_RED as GL_DEPTH_TEXTURE_MODE when GL_ARB_texture_rg is
supported.  This was done as a separate test so that when an
implementation add support for GL_ARB_texture_rg the status of
depth-tex-modes won't change, but the status of depth-tex-modes-rg may
change from 'skip' to 'fail' or 'pass'.