glsl-reload-source: use gl_Vertex instead of generic position attribute
Binding the "position" attribute to 0 and using the conventional
GL_VERTEX_ARRAY array (in piglit_draw_rect()) is a fuzzy area in the
spec.  It doesn't work in Mesa; not sure about other OpenGL implementations.

Since the point of this piglit test is to test shader source
reloading, this is a separate issue that should be exercised in
another test, if we care about it.