Add shaders from GStreamer as parser tests.
These were hand-copied from commit
7edf5771853fa34099b07de1c2b18e03db9048dd of gst-plugins-gl.  C strings
were converted to standalone files, with adjustment of whitespace.
All changes other than string conversion and whitespace are annotated
with a comment mentioning piglit.

Mesa fails on the array constructors for 3 of them, while GLSL2 fails on many
of them due to lack of builtin functions but not due to array constructors.