glsl-deadcode-call: Fix GLSL compilation error.
The NVIDIA, fglrx, and Apple GLSL compilers all failed to compile this
shader because the function func, which has a return type of float, did
not return anything. The Mesa GLSL compiler, however, did succesfully
compile this shader.

A separate negative test case should be written to test that functions
with return types that don't return values do not pass parsing.