trinity-fp1: fix invalid glMultiTexCoord2f() calls
The first parameter isn't an integer but a GL_TEXTUREx token.
Because Mesa masks the tex unit param with 0x7 we were getting the
right texture unit anyway but that might not work with other OpenGL
implementations.

Also, don't use texcoord S=1.0 with GL_REPEAT mode since a little
bit of arithmetic error (such as with SSE) can cause this to become
integer texcoord 0 or 255 (if width=256) depending on the multiply
instruction.  Use zero to be more reliable.

Fixes gallium/softpipe/sse failure (bug 23318).