glsl-orangebook-ch06-bump: Try to avoid depending on quad-to-tris behavior.
I noticed this when the single quad to trifan change on 965 broke this
testcase slightly.  The quad we're drawing is so large that the light
and eye direction vectors change massively over the surface, and
interpolation of them gives unintuitive results.  If the
implementation selected (0,1,2,3) as the order for the trifan instead
of (3,0,1,2), the specular highlight appeared as a bright line from
bottom left to top right, instead of there being an edge to the
specular from top left to bottom right.

Since GL doesn't spec a specific way to rasterize quads when breaking
them down to tris (see page 165 of the GL 3.2 compat profile spec),
fix the testcase to make quad order less significant, and use
the new quad-invariance case for testing the order.