glsl-fs-texture2d*-bias*: Tests for texture2D behavior with an LOD bias.
These simple tests don't make sure that the biasing came out right
(the sample texture has the same image at all levels), but it is
useful in terms of making sure the rest of the texture handling
worked.  The focus here is on the different function signatures of
texture2D with bias, as opposed to glsl-lod-bias which tests the
actual biasing.