fbo-drawbuffers-maxtargets: Gracefully handle glUseProgram errors.
On Mac OS X the fragment shader doesn't compile which eventually results
in an assertion failure in the test. Check for glUseProgram error
earlier and gracefully report a failure instead.

The Apple GLSL compiler gives this compilation error.
Failed to compile FS: ERROR: 0:4: 'gl_FragData' : Indirect access to
fragment shader output is not handled
WARNING: 0:4: 'gl_FragData[0]' : Indirect reference to an array with a
single element, not likely the desired construct.