Add missing "#version 120" to tests expecting implicit int->float conversion
According to the latest GLSL specification, code without a #version
directive will be assumed to target GLSL version 1.10. And on the
topic of type conversion, the GLSL 1.10 specification says:

	The assignment operator stores the value of expression into
	lvalue. It will compile only if expression and lvalue have the
	same type. All desired type-conversions must be specified
	explicitly via a constructor. [Section 5.8]

Meanwhile, these three tests each have a construct in which an integer
rvalue is assigned to a float lvalue. This implicit conversion is
supported as of GLSL 1.20, so adding a "#version 120" directive makes
these tests conformant.