Add valgrind-based testing.
This works by automatically constructing a valgrind group of tests
which includes every existing test. These tests run the existing test,
but inside valgrind and pass only if the underlying test passes and
valgrind detects no errors.

If the underlying test passes and valgrind detects an error then the
valgrind-based test reports a failure.

If the underlying test fails then the valgrind-based tests reports a
skip.

The valgrind-based tests are ssomewhat slow for obvious reasons, so
they can be skipped by taking advantage of the new -x|--exclude-tests
option, by passing an argument to piglit-run.py such as:

	--exclude-tests=valgrind