Add more redeclaration tests; fix redeclaration-02 and struct-05.
In 1.10, variables and functions have separate name spaces, so both can
be declared in the same scope, and variables do not hide functions.
This means that redeclaration-02.vert should pass, not fail.  However,
structures conflict with both variables and functions, so struct-05.vert
should fail.

In 1.20, functions, types, and variables all share the same namespace.
Thus, declaring any two in the same scope is an error (aside from
function overloading).  Also, declaring a name hides declarations of
_any_ kind in an outer scope.

Issues 22-28 in the GLSL 1.20 specification shed a lot of light on
the scoping rules (though it's still not 100% clear).

Also, according to the 1.20 spec, built-ins live in a special scope
beyond a shader's global scope.