glsl-uniform-out-of-bounds: Make this test access the end of b[3].
It was relying on Mesa's lack of trimming of uniform length (which is
required by spec) so that b[2] existed and had a location.  It's not
clear to me what the desired testing was here (only b[0] being used
but c[1..4] being used suggested that b[] was supposed to be testing
this behavior, but b[] was the only one getting the array offset
testing when it should only have succeeded on c[] suggests not).  This
is the minimal change to keep the test testing the same aspects of the
API it was before.