Remove CorrectPreprocess11.frag test from glslparser.tests
This test includes some bogus code, (redefining a macro to a different
value and not expecting any error from that). The C99 specification is
quite clear (a macro "shall not" be redefined unless equivalent),
while the GLSL specification is very vague (#define to be implmented
"as is standard for C++ compilers").

A survey of other implementations suggest that some do just generate a
warning rather than an error for this case. But this test really
should be fixed to not rely on the behavior of re-defined macros I
think.