pointAtten: improve measureSize() function
Before, we read one row of pixels where we expected the point to be
in order to compute the point size.  With some drivers single pixel
points might be positioned one above or below the expected position
thus causing the test to fail.  Now, read thee rows of pixels to make
sure we don't miss the point.  The purpose of this test is to measure
attenuated point size, not exact positioning of single pixel points.