framework: Add class GLSLParserTest
This test class feeds a GLSL source file to the glslparsertest executable.
Information about the test (such as expected result and GL requirments) may be
placed directly in the shader source file as a specially formatted comment
block.

Also, modify PlainExecTest to skip tests for which 'command == None'.  This
prevents glslparsertests that lack a config section from failing in odd ways.