all.tests: Modify ValgrindExecTest to support GLSLParserTest
Summary
-------
ValgrindExecTest made an assumption about the test which it wraps that did not
hold for GLSLParserTest. This patch introduces a workaround hack in
ValgrindExecTest that eliminates that assumption.

Problem Details
---------------
When the ValgrindExecTest initializer wraps a pre-existing PlainExecTest, it
made the assumption that the PlainExecTest had already set its 'command'
field. This does not hold true for GLSLParserTest,  because GLSLParserTest
cannot construct the command until it reads the test file, which is not done
at initialization. (GLSLParserTest postpones the construction of the 'command'
filed in order to avoid the penalty of opening and reading test files that
will not be ran.)

However, even if GLSLParserTest /did/ attempt to construct the 'command' field
at initialization, sometimes the construction fails due to an incorrectly
formatted test file. In this case, there is no sensible value for the
'command' field.

Fix Details
-----------
ValgrindExecTest postpones reading the wrapped test's 'command' field until
ValgrindExecTest.run(). If the 'command' field is None, it reports test
failure.