Correct errors in the fbo-incomplete-texture tests
These tests originally wanted the FBO to be incomplete.  However, this
just verified incorrect behavior in another vendor's driver.  The spec
says that all of these cases should produce a complete FBO.

The format of the texture was also change from RGB to RGBA.  This
should prevent FRAMEBUFFER_UNSUPPORTED on some drivers.