cmake: Each CMakeLists.txt that produces a target, rename to CMakeLists.gl.txt
In preparation for adding GLES2 targets to Piglit, it is necessary to
compartmentalize the GL targets and the GLES2 targets (of which none yet
exist) in API-specific CMake files. Since all CMake targets currently use
the OpenGL API, achieving compartmentalization is as simple as renaming
each CMakeLists.txt.

Note: This belongs to a series that adds GLES2 support to Piglit.
Signed-off-by: Chad Versace <chad.versace@intel.com>