glx/pixmap-life: New test
The spec, though a little poorly worded, seems to only have one
interpretation about pixmap lifetime.  Namely, that GLX pixmaps take a
reference on the underlying X pixmap, and that the storage for a GLX
pixmap is not deleted until the GLX pixmap itself is destroyed (even if
the association between the Pixmap XID and the pixmap is torn down by
calling XFreePixmap).

Verify this by deleting a GLX pixmap in both orders and barking if
there's any error.

Signed-off-by: Adam Jackson <ajax@redhat.com>