glx-fbconfig-sanity: Relax (remove) checks on GLX_PIXMAP_BIT
GLX 1.3 does not require that fbconfigs with GLX_PIXMAP_BIT set have an
associated visual.  The extent of its opinion on the matter of
GLX_PIXMAP_BIT is that an fbconfig with that bit set may be used to
create GLXPixmaps.  The extent of its opinion on GLX_VISUAL_ID is that
if an fbconfig supports windows then GLX_VISUAL_ID will be non-zero.
Actually, it says that GLX_VISUAL_ID will be set to "the XID of the
Visual".  This is an understandable mistake, since the sample server has
always allocated Visual IDs from the XID allocator, even though the
X11 protocol spec declares them to be independent namespaces.

At any rate, there are more pixmaps in heaven and earth, Horatio, than
are dreamt of in your visuals.