glx-fbconfig-sanity: Be more strict about GLX_DRAWABLE_TYPE
The spec doesn't say what you could possibly do with an fbconfig whose
GLX_DRAWABLE_TYPE value is 0, but it doesn't say it's illegal either.
Warn if there's not already been a failure, since it's going to be very
difficult to use such an fbconfig and apps may well get confused.