glsl-fs-clamp-[234]: New tests for a bug Chad noticed in saturate handling.
The point here is to try to catch the cases where the compiler turns a
min/max of a variable float with constant vectors into a saturate of a
float.

We never actually failed, becase we aren't writemasking in ir_to_mesa
(so the full temporary vector would be filled anyway), and because in
brw_fs.cpp we had split up the expression to scalars already.