glsl-1.20: Add a bunch of outerProduct execution tests for constant data
The previous batch tested non-constant parameters to outerProduct.
All of those shaders should execute the outerProduct code on the GPU.
This batch adds the same set of tests but using constant parameters to
outerProduct.  These are likely executed in the compiler on the CPU
during a constant folding pass.

Testing all of the possible paths seems to make sense.