glslparsertest: Also call glLinkShader on shaders that compiled.
This doesn't impact the piglit results (since the shaders are not
categorized according to whether they're expected to link, and many
aren't), unless glLinkShader happens to cause a crash.  However, doing
this increases coverage of the shader linking code, and also means
that glslparsertest can be used in analysis using shader-db.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Chad Versace <chad@chad-versace.us>