glx-tfp: Use XRender to correctly set up alpha channel.
The previous way of initializing the "alpha" channel of the pixmap was
to just set the pixel bits other than r, g, b, to 1.0 or 0.5 times
their maximum value.  However, thanks to the delights of xlib, a pixel
value is a long, not a sized integer.  So, when we set alpha to half
of ~(r, g, b,), on 64-bit that was half of 0xffffffffff000000 instead
of half of 0xff000000, which resulted in black.

XRender is the correct X rendering API for working with alpha
channels, so use that instead.  The ugly bit here is that to preserve
the existing test behavior, we have one of the colors filled be (1.0,
0.0, 0.0, 0.5), which is non-premultiplied and thus not really proper
Render rendering.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=24491