hiz: Fix hiz-depth-test-* to check just depth test and not depth reads
The hiz-depth-test-* tests were not really *unit* tests. They tested both
depth tests and depth reads.

Now, they probe only color buffer and not the depth buffer. They should
not not probe the depth buffer because correct operation of 1) depth
testing and depth writes (via glDraw*) and of 2) depth buffer reads (via
glRead*) are independent. It is possible for 1 to work while 2 to fail.
The tests now cover only case 1.

The distinction here is not pedantic. On Intel SandyBridge, with hiz only
partially implemented, case 1 passes and case 2 fails.

Signed-off-by: Chad Versace <chad@chad-versace.us>