glsl-1.10: Add some tests for over-riding built-in functions
These are somewhat justified by the GLSL spec:

    "A shader can redefine built-in functions. If a built-in function
    is redeclared in a shader (i.e. a prototype is visible) before a
    call to it, then the linker will only attempt to resolve that call
    within the set shaders that are linked with it."

This also matches the behavior that the OpenGL Shading Language (the
"orange" book) tells people to expect:

    "If the definition is in a different shader, just make sure a
    prototype is visible before calling the function. Otherwise, the
    built-in version is used."

All of these tests pass on NVIDIA's closed-source Linux driver.

All of these tests pass on AMD's closed-source Linux driver except
override-builtin-*-07.

All of these tests pass on Mac OS 10.6.8.

Many of these tests fail on Mesa: override-builtin-const-0[57] and
override-builtin-uniform-0[13567].

I have not tested Intel's Windows driver.

v2: Adds two more tests (override-builtin-0[56]) suggested by Ken and
Chad.  It also adds versions of all six tests that use uniforms as
parameters to the built-in instead of constants.

v3: Add several more tests (override-builtin-1[3456]) that are
variations of the existing glsl-override-builtin-2 test.  Renamed the
tests so that the constant tests have 'const' in the name and the
non-constant tests have 'uniform' in the name.  The tests with the
same number should be essentially identical.

This is related to Mesa bug #31744.

Tested-by: Vinson Lee <vlee@vmware.com>