fbo-stencil: assorted fixes to account for rasterization variance
Since the window size is odd (123 x 123), when we draw polygons over
the four quadrants the middle row/column of pixels may be touched by
different polygons depending on minor differences in transformation
and rasterization.  Simply ignore the middle/row column when verifying
the results.

For the copypixels case, adjust the source x/y coords to skip the
middle row/column of pixels to avoid the variance mentioned above.

Finally, in the error message printf, prefix values with "0x" to
indicate that they're hexadecimal.