texunits: Fix the texture env tests.
There's nothing in the spec that restricts TexEnv* to fixed function
texture units. In addition, MAX_TEXTURE_IMAGE_UNITS is the number of
fragment units, not the number of fixed function units. This was failing
on fglrx, and with Mesa since bfe284fd26e96b71c7cf46e6365b3697d68cde83.
With the nvidia blob this was mostly skipped due to
MAX_TEXTURE_IMAGE_UNITS being equal to MAX_COMBINED_TEXTURE_IMAGE_UNITS.
This is for fdo bug #39074.

Signed-off-by: Henri Verbeet <hverbeet@gmail.com>
Reviewed-by: Brian Paul <brianp@vmware.com>