Fix recursion in add_shader_test_dir().
When attempting to recursively add subdirectories,
add_shader_test_dir() was using the basename, rather than the full
path, to determine whether a directory entry was a subdirectory.

Fortunately, this didn't cause any shader tests to be skipped, because
none of the directories it was attempting to visit recursively
actually contained any subdirectories.

Reviewed-by: Chad Versace <chad@chad-versace.us>
Tested-by: Ian Romanick <ian.d.romanick@intel.com>