glsl-1.10 / glsl-1.20: Add non-constant index write tests
All of these tests follow a similar form.  An array of matrices is set
to some values that have one incorrect element.  A non-constant index
write is used to fix that incorrect element.  The result is verified.
The tests vary on:

  - Storage qualifier of the array/matrix (uniform, temporary, varying).
  - Array / not an array.
  - Matrix size (mat2, mat3, mat4).
  - Whether or not the array index is a non-constant.
  - Whether or not the matrix column is a non-constant.
  - Whether or not the matrix row is a non-constant.

Since the base type is always a matrix, attributes and fragment shader
outputs are not tested.