framework: Write each test result to the result file as the test completes
When a test run is interrupted, perhaps by a system crash, we often want
the test results. To accomplish this, Piglit must write each test result
to the result file as the test completes.

If the test run is interrupted, the result file will be corrupt. This is
corrected in a subsequent commit.

CC: Ian Romanick <ian.d.romanick@intel.com>
Signed-off-by: Chad Versace <chad@chad-versace.us>