framework: Repair result file if result file is incomplete or corrupt
If the JSON result file was not closed properly, perhaps due a system
crash during a test run, then TestrunResult.parseFile will attempt to
repair the file before parsing it.

The repair is performed on a string buffer, and the result file is never
written to. This allows the result file to be safely read during a test
run.

CC: Ian Romanick <ian.d.romanick@intel.com>
Signed-off-by: Chad Versace <chad@chad-versace.us>