Add comprehensive tests of builtin functions with uniform input.
This patch adds 462 tests to piglit, which comprehensively test the
behavior of GLSL built-in functions on vertex and fragment shaders.
The test vectors are auto-generated in Python, by using the numpy
library to simulate the behavior of the built-in functions on a
variety of inputs (in builtin_function.py), and then producing
shader_runner tests to verify the expected behavior (in
gen_builtin_uniform_tests.py).

Test generation happens as part of the piglit build process.  To run
the test generation step by itself, use "make gen-tests" from the root
of the piglit tree (after configuring using cmake).  The generated
tests may be found in generated_tests/spec.

The auto-generation code depends on Python 2.6 and on the numpy
library (www.numpy.org).  I've updated the README to reflect these
dependencies, and added code to CMakeLists.txt to check for them at
configure time.  (Note: other parts of Piglit already depend on Python
2.6, but the dependency wasn't previously explicit).