Add constant array size builtin tests.
This patch adds a new set of generated tests which parallels those
introduced by commt cccc419e (Add comprehensive tests of builtin
functions with uniform input.)

The new tests validate the correctness of built-in functions when used
to compute array sizes.  The tests work by creating arrays whose size
is 1 if the result of applying the built-in function is within
tolerance of the correct answer, and -1 if not.  Since negative array
sizes are prohibited, any improperly computed values will generate a
compile error.

Though it may seem absurd to compute an array size based on the result
of calling a built-in function, it's worth testing for two reasons:

(1) It is explicitly allowed in GLSL versions since 1.20.

(2) Since array sizes need to be computed at compile time, this is a
convenient way of testing that the GLSL compiler properly implements
constant folding of built-in functions.  Indeed, the original
motivation for these tests was to validate bug fixes to Mesa's
constant folding logic.

Here is an example of one of the generated tests (this test is
exp2-vec2.vert):

/* [config]
 * expect_result: pass
 * glsl_version: 1.20
 * [end config]
 *
 * Check that the following test vectors are constant folded correctly:
 * exp2(vec2(-2.0, -0.66666669)) => vec2(0.25, 0.62996054)
 * exp2(vec2(0.66666669, 2.0)) => vec2(1.587401, 4.0)
 */

void main()
{
  float[distance(exp2(vec2(-2.0, -0.66666669)), vec2(0.25, 0.62996054)) <= 6.7775386e-06 ? 1 : -1] array0;
  float[distance(exp2(vec2(0.66666669, 2.0)), vec2(1.587401, 4.0)) <= 4.3034684e-05 ? 1 : -1] array1;
  gl_Position = vec4(array0.length() + array1.length());
}