Refactor PlainExecTest, GleanTest, and GTFTest to share code.
Each of these classes are essentially the same, except for:
1. How to determine whether the test passed/failed
2. Setting up the command and options

All the logic for running the command and obtaining its output can be
shared, as well as the logic to determine if it trapped/aborted/etc.

This patch creates a new 'ExecTest' base class.  Subclasses can set the
exact command to run and options in their __init__ method, and must
supply an 'interpretResult' method which reads the output and sets
results['result'] to pass/fail (along with other desired annotations).

v2: Update for JSON related changes and ensure that output stays the
    same.  (Piglit and Glean are identical; GTF now includes the command
    that was run but is otherwise equivalent.)

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>