Test implicit type conversion of out parameters.
This patch adds six tests of the proper behavior of function out
parameters when used in a context that requires implicit type
conversion.

Two of the tests are based on an explicit example in the GLSL-1.30
spec.  The remaining tests adapt this example to GLSL 1.20 (which
seems to be intended to have the same conversion rules, though they
are less clearly articulated).