Test interaction between globals and functions pulled in by linking.
This test exercises a bug in Mesa.  If a function that is not known to
the compilation unit containing main() (IOW, neither defined not
declared in that compilation unit) attempts to access a global
variable, then improper IR is generated.

I'll be submitting a patch to Mesa to fix this problem, and to check
the IR validity after linking.  This test will ensure that the fix
won't regress.

Reviewed-by: Eric Anholt <eric@anholt.net>