Changed piglit-summary-html to explicitly encode its output as UTF-8.
Previously piglit-summary-html didn't explicitly encode its output at
all--it relied on the default behavior of the Python file.write()
method (which, in Python 2.x, is to write 8-bit strings verbatim and
to encode unicode strings using ASCII).  This meant that if a
non-ASCII character appeared in any of the test data,
piglit-summary-html would fail with an exception.

This patch changes piglit-summary-html to explicitly encode its output
using UTF-8.  It also updates all of the HTML templates to specify
UTF-8 as the encoding, so that when results are viewed in a web
browser, non-ASCII characters will display correctly.