Remove bogus vs-clip-distance-implicitly-sized.shader_test.
This test was attempting to verify that the GLSL compiler properly
computed the size of the gl_ClipDistance array when it was implicitly
sized, by calling .length() on the array.  But according to the GLSL
spec, it is illegal to call .length() on an implicitly sized array.

Because of this, it is actually impossible to tell using a test
whether the GLSL compiler properly computed the size of an
implicitly-sized array; all we can do is attempt to use all elements
of the array and make sure nothing breaks.  The test
vs-clip-distance-all-planes-enabled.shader_test already does this.  So
I've removed the bogus test.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>