Add --check-link option to glslparsertest.
This option tells glslparsertest to treat link errors the same as
compile errors.  For tests that expect a result of "pass", this causes
glslparsertest to verify that the shader both compiles and links
without error.  For tests that expect a result of "fail", this causes
glslparsertest to treat a link-time error as passing the test, just as
it would a compile-time error.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>