Fragment shader tests for clipDistance sizing.
This patch adds two new tests,
fs-clip-distance-explicitly-sized.shader_test and
fs-clip-distance-sizeable-to-max.shader_test, to verify that the
clipDistance array behaves properly when its size is explicitly set in
the fragment shader.

These tests parallel the existing tests,
vs-clip-distance-explicitly-sized.shader_test and
vs-clip-distance-sizeable-to-max.shader_test, which were testing the
equivalent functionality in the vertex shader.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>