Add tests of whole-array assignment and out/inout params.
In GLSL 1.10, whole arrays are not lvalues (even though their elements
are).  This means it is illegal to use an array on the left hand side
of an assigment, or to use it as an out or inout parameter of a
function.

In GLSL 1.20 and later versions, an array is a proper lvalue, so all
of these uses are allowed.

In GLSL 1.00 ES, it is illegal to use an array on the left hand side
of an assignment, but using it as an out or inout parameter of a
function is ok.

This patch adds tests to verify correct behavior in GLSL 1.10, GLSL
1.20, and GLSL ES 1.00.  In addition it verifies that in GLSL 1.20, a
built-in array may be used as an lvalue (since this currently
exercises a different code path in Mesa).

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>