glslparsertest: use dummy shaders when linking for GLSL 1.00 ES.
GLES requires both vertex and fragment shaders to be present in order
to link.  From section 2.10.3 (Program Objects) of the GLES 2.0 spec:

    "Linking will also fail ... if program does not contain both a
    vertex shader and a fragment shader ..."

This patch adds logic to glslparsertest so that when attempting to
verify that a vertex program links, it attaches a dummy fragment
program, and vice versa.  This logic is only used when the requested
GLSL version is 1.00 (indicating that GLES should be used).

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>