Test "double negative" corner case in copy propagation.
This patch adds two tests that exercise a subtle corner case bug found
in Mesa with i965: If a value is negated, and then negated again, then
copy propagation would fail to un-do the negation.

The bug only exists on i965 vertex shaders, but I went ahead and
added the equivalent fragment shader test for completeness.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>