Test use of gl_ClipDistance with uint indices.
Because of the way the gl_ClipDistance array is handled in Mesa,
accesses using a uint index may not take precisely the same code path
as accesses using an int index.  This test verifies that
gl_ClipDistance works properly when it is accessed using a uint index.

Tested using the nVidia proprietary Linux driver.