Update clip distance tests to use 8 clip distances.
The clipping tests previously in Piglit only tested the first 6
clipping planes, since prior to GLSL 1.30, a conforming implementation
was allowed to only support 6 clipping planes.  However, in GLSL 1.30,
the minimum value of gl_MaxClipDistances and gl_MaxClipPlanes was
increased to 8.

This patch modifies vs-clip-distance-all-planes-enabled.shader_test to
exercise all 8 clipping planes.

The previous implementation of
vs-clip-distance-all-planes-enabled.shader_test used to also verify,
as a side effect, that gl_ClipDistance could be implicitly sized.  The
new version does not (it requires gl_ClipDistance to be explicitly
sized, since it indexes into it using a loop variable).  So I've added
a new test, vs-clip-distance-implicitly-sized.shader_test, to verify
this.

Tested using the nVidia proprietary Linux driver.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>