Test proper behavior of gl_MaxClipDistances and gl_MaxClipPlanes.
This patch adds a test to verify that if the implementation reports
support for GLSL version 1.30, the values of gl_MaxClipDistances and
gl_MaxClipPlanes (accessible via GLSL) match the values of
MAX_CLIP_DISTANCES and MAX_CLIP_PLANES queried using glGetIntegerv.

Furthermore, it checks that the value of MAX_CLIP_DISTANCES matches
the value of MAX_CLIP_PLANES (since the latter is a deprecated name
for the former), and that the value is at least 8 (since GLSL 1.30
requires at least 8 clip distances to be supported).

Tested using the nVidia proprietary Linux driver.