Test that clip planes behave properly under coordinate transformation.
This test measures the effect on clip planes of:
- model-view transformations at the time clip planes are specified
- projection transformations at the time clip planes are specified
- model-view transformations at the time of drawing
- projection transformations at the time of drawing
- transformations between gl_Vertex and gl_Position
- transformations between gl_Vertex and gl_ClipVertex

The results are compared with those predicted from GLSL spec versions
1.10, 1.20, and 1.30 and OpenGL 2.1 and 3.0.  All of these specs
agree, except that the GLSL 1.30 spec dictates what the behavior
should be when the vertex shader sets gl_Position but not
gl_ClipVertex; all the other specs leave this behavior undefined.

Tested using the nVidia proprietary Linux driver, which passes in all
cases except when the vertex shader sets gl_Position but not
gl_ClipVertex (a known issue with this driver).