Test that gl_ClipVertex respects clip plane enable flags.
This test is similar to the existing
"vs-clip-distance-enables.shader_test", except that it tests the
behavior of gl_ClipVertex instead of gl_ClipDistance.  Because of
differences in the way gl_ClipVertex and gl_ClipDistance are
implemented in Mesa, it's possible that one of these tests might pass
while the other fails.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>