Add a test of isinf() and isnan() functionality.
isinf() and isnan() are difficult to test, since the GLSL 1.30 spec
does not define when an implemenation is required to generate infinite
or NaN values; in fact, it explicitly allows for implementations that
do not even have a representation of infinity or NaN.

This test verifies that if the implementation does generate infinite
or NaN values, they behave in a consistent way, and that when it
claims that a value is finite, it truly is finite.

Tested on the nVidia proprietary Linux driver, which passes all tests.

Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>