fbo-blit-d24s8: add GLUT_DEPTH to piglit_window_mode
This test tries to write to the window's depth buffer, but we weren't
requesting one when we created the window.

This fixes a failing glGetError()==0 assertion but we still generate
an INVALID_OPERATION error with swrast because we're trying to blit
from an FBO with 24-bit Z values to the window with 16-bit Z values.
That may be a bogus Mesa error check.

Reviewed-by: Eric Anholt <eric@anholt.net>