Test that fixed vertex functionality respects clip plane enable flags.
This test reproduces bug 41603
(https://bugs.freedesktop.org/show_bug.cgi?id=41603).  It is similar
to the existing "vs-clip-vertex-enables.shader_test", except that it
tests the behavior of fixed vertex functionality rather than a GLSL
vertex shader.

Note: since this test is checking for a bug in fixed vertex
functionality, strictly speaking it doesn't belong in the GLSL-1.20
directory.  But it seemed better to put it here with other related
clipping tests than to put it by itself in some other directory.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>