glsl-sin, glsl-cos: move piglit_ortho_projection() calls to display() func
On Windows, if the initial window size is too small it gets bumped up.
This upsets the ortho projection, which depends on the window size.
Call piglit_ortho_projection() in the display() func when we know the
real window size.

Fixes bogus test failures on Windows.

See commit 3a18e84dba8cd94eeab4ec13e1848cbb7c3cb953 and the following
commits for other instances of this fix.