Make Python scripts independent of the current working directory
The main purpose of this patch is to make piglit independent of the
current working directory, so it is possible to package piglit as a RPM
package (with binaries symlinked to /usr/bin, most of the files in
read-only /usr/lib*/piglit directory, and results somewhere else).

So it is now possible to run

$ piglit-run tests/quick-driver.tests /tmp/piglit

and then with this command

$ piglit-summary-html --overwrite /tmp/piglit/results /tmp/piglit/main

generate a report in /tmp/piglit/results/index.html & al.

Signed-off-by: Matěj Cepl <mcepl@redhat.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>