Flip the expected result for 'GLSL link matched global initializer expression'
This test, formerly known as glsl-link-initializer-05, used to verify
that two shaders that initialized a global variable with the same
non-constant expression would succeed linking.  However, this did not
match the behavior of at least one other major vendor.  In addition,
the language about initializers changed in GLSL 4.20.

    "If a shared global has multiple initializers, the initializers
    must all be constant expressions, and they must all have the same
    value. Otherwise, a link error will result. (A shared global
    having only one initializer does not require that initializer to
    be a constant expression.)"

The test now expects the GLSL 4.20 behavior.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Chad Versace <chad@chad-versace.us>
Acked-by: Paul Berry <stereotype441@gmail.com>